fix last hour bug, fix issue #53
This reverts most of the solution proposed by the previous commit.

I used a slightly hackish method to fix fatal errors caused by
the lack of a 'next hour' for the 'last hour' of the day in GMT
when trying to determine whether temperature is rising or falling.
I just assumed that the temperature between 18:00 and 19:00 EST will
usually be falling...